fix future exception for dropshipping
In a previous commit by yours truly, I disabled the stock-based checks
in dropshipping situations, as was suggested in #47.

The idea was correct, but the method future_orders_are_affected needs to
return False to skip the check (it means that future orders are _not_
affected, and we are cool).